Crate markup5ever

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Reexports

pub extern crate tendril;

pub use interface::ExpandedName;

pub use interface::QualName;

pub use interface::Attribute;

Modules

buffer_queue 	 [
Experimental
]

data 	 [
Experimental
]

interface 	 [
Experimental
]

rcdom 	 [
Experimental
]
A simple reference-counted DOM.

serialize 	 [
Experimental
]

smallcharset 	 [
Experimental
]

Macros

expanded_name 	 [
Experimental
]

local_name 	 [
Experimental
]

namespace_prefix 	 [
Experimental
]

namespace_url 	 [
Experimental
]

ns 	 [
Experimental
]

small_char_set 	 [
Experimental
]

Structs

LocalNameStaticSet 	 [
Experimental
]

NamespaceStaticSet 	 [
Experimental
]

PrefixStaticSet 	 [
Experimental
]

SmallCharSet 	 [
Experimental
]
Represents a set of "small characters", those with Unicode scalar values less than 64.

Type Definitions

LocalName 	 [
Experimental
]

Namespace 	 [
Experimental
]

Prefix 	 [
Experimental
]