Module markup5ever::interface::tree_builder

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Reexports

pub use self::NodeOrText::AppendNode;

pub use self::NodeOrText::AppendText;

pub use self::QuirksMode::Quirks;

pub use self::QuirksMode::LimitedQuirks;

pub use self::QuirksMode::NoQuirks;

Structs

ElementFlags 	 [
Experimental
]

Enums

NextParserState 	 [
Experimental
]
Whether to interrupt further parsing of the current input until the next explicit resumption of the tokenizer, or continue without any interruption.

NodeOrText 	 [
Experimental
]
Something which can be inserted into the DOM.

QuirksMode 	 [
Experimental
]
A document's quirks mode.

Traits

Tracer 	 [
Experimental
]
Trace hooks for a garbage-collected DOM.

TreeSink 	 [
Experimental
]

Functions

create_element 	 [
Experimental
]