Struct markup5ever::rcdom::RcDom

pub struct RcDom {
    pub document: Handle,
    pub errors: Vec<Cow<'static, str>>,
    pub quirks_mode: QuirksMode,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The DOM itself; the result of parsing.

Fields
document: Handle

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The Document itself.

errors: Vec<Cow<'static, str>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Errors that occurred during parsing.

quirks_mode: QuirksMode

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The document's quirks mode.