Type Definition markup5ever::rcdom::Handle

type Handle = Rc<Node>;

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Reference to a DOM node.