markup5ever::serialize

Trait Serialize

pub trait Serialize {
    // Required method
    fn serialize<S>(
        &self,
        serializer: &mut S,
        traversal_scope: TraversalScope,
    ) -> Result<()>
       where S: Serializer;
}

Types that can be serialized (according to the xml-like scheme in Serializer) implement this trait.

Required Methods

fn serialize<S>( &self, serializer: &mut S, traversal_scope: TraversalScope, ) -> Result<()>
where S: Serializer,

Take the serializer and call its methods to serialize this type. The type will dictate which methods are called and with what parameters.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.