Trait markup5ever::serialize::Serializer

pub trait Serializer {
    // Required methods
    fn start_elem<'a, AttrIter>(
        &mut self,
        name: QualName,
        attrs: AttrIter
    ) -> Result<()>
       where AttrIter: Iterator<Item = AttrRef<'a>>;
    fn end_elem(&mut self, name: QualName) -> Result<()>;
    fn write_text(&mut self, text: &str) -> Result<()>;
    fn write_comment(&mut self, text: &str) -> Result<()>;
    fn write_doctype(&mut self, name: &str) -> Result<()>;
    fn write_processing_instruction(
        &mut self,
        target: &str,
        data: &str
    ) -> Result<()>;
}

Types that are capable of serializing implement this trait

Required Methods

fn start_elem<'a, AttrIter>( &mut self, name: QualName, attrs: AttrIter ) -> Result<()>where AttrIter: Iterator<Item = AttrRef<'a>>,

Serialize the start of an element, for example <div class="test">.

fn end_elem(&mut self, name: QualName) -> Result<()>

Serialize the end of an element, for example </div>.

fn write_text(&mut self, text: &str) -> Result<()>

Serialize a plain text node.

fn write_comment(&mut self, text: &str) -> Result<()>

Serialize a comment node, for example <!-- comment -->.

fn write_doctype(&mut self, name: &str) -> Result<()>

Serialize a doctype node, for example <!doctype html>.

fn write_processing_instruction( &mut self, target: &str, data: &str ) -> Result<()>

Serialize a processing instruction node, for example <?xml-stylesheet type="text/xsl" href="style.xsl"?>.