Type Alias markup5ever::serialize::AttrRef

pub type AttrRef<'a> = (&'a QualName, &'a str);

A type alias for an attribute name and value (e.g. the class="test" in <div class="test"> is represented as (<QualName of type class>, "test").

This is used in Serializer::start_elem where the value being serialized must supply an iterator over the attributes for the current element