Crate markup5ever

Re-exports

pub use interface::Attribute;

pub use interface::ExpandedName;

pub use interface::QualName;

pub use interface::TokenizerResult;

pub use tendril;

Modules

buffer_queue

The BufferQueue struct and helper types.

data

interface

Types for tag and attribute names, and tree-builder functionality.

serialize

Traits for serializing elements.

smallcharset

This module contains a single struct SmallCharSet. See its documentation for details.

Macros

expanded_name

Helper to quickly create an expanded name.

local_name

Takes a local name as a string and returns its key in the string cache.

namespace_prefix

Takes a namespace prefix string and returns its key in a string cache.

namespace_url

Takes a namespace url string and returns its key in a string cache.

ns

Maps the input of namespace_prefix! to the output of namespace_url!.

small_char_set

Create a SmallCharSet, with each space-separated number stored in the set.

Structs

LocalNameStaticSet

NamespaceStaticSet

PrefixStaticSet

SmallCharSet

Represents a set of “small characters”, those with Unicode scalar values less than 64.

Type Aliases

LocalName

Namespace

Prefix