Module markup5ever::interface::tree_builder

This module contains functionality for managing the DOM, including adding/removing nodes.

It can be used by a parser to create the DOM graph structure in memory.

Re-exports

pub use self::NodeOrText::AppendNode;

pub use self::NodeOrText::AppendText;

pub use self::QuirksMode::LimitedQuirks;

pub use self::QuirksMode::NoQuirks;

pub use self::QuirksMode::Quirks;

Structs

ElementFlags

Special properties of an element, useful for tagging elements with this information.

Enums

NextParserState

Whether to interrupt further parsing of the current input until the next explicit resumption of the tokenizer, or continue without any interruption.

NodeOrText

Something which can be inserted into the DOM.

QuirksMode

A document’s quirks mode, for compatibility with old browsers. See quirks mode on wikipedia for more information.

Traits

Tracer

Trace hooks for a garbage-collected DOM.

TreeSink

Methods a parser can use to create the DOM. The DOM provider implements this trait.

Functions

create_element

A constructor for an element.