Module markup5ever::serialize

Traits for serializing elements. The serializer expects the data to be xml-like (with a name, and optional children, attrs, text, comments, doctypes, and processing instructions). It uses the visitor pattern, where the serializer and the serializable objects are decoupled and implement their own traits.

Enums

TraversalScope
Used as a parameter to serialize, telling it if we want to skip the parent.

Traits

Serialize
Types that can be serialized (according to the xml-like scheme in Serializer) implement this trait.

Serializer
Types that are capable of serializing implement this trait

Type Definitions

AttrRef
A type alias for an attribute name and value (e.g. the class="test" in <div class="test"> is represented as (<QualName of type class>, "test").